requirements-dev.txt
httpie==0.9.1
django-debug-toolbar
